replace casts long to int with Math.toIntExact (#1072)
* Replace unsafe (int) long casts with Math.toIntExact()

Replace explicit (int) casts of long values with Math.toIntExact() to
detect integer overflow at runtime rather than silently truncating.

Changes cover:
- hemf emf records: HemfComment, HemfFill, HemfPalette, HemfDraw, HemfFont
- hpsf: Section, Array, Vector, Property, PropertySet, PropertySetFactory, VariantSupport
- poifs/crypt: CryptoAPIDecryptor, ChunkedCipherOutputStream, XORDecryptor, XOREncryptor
- poi-ooxml XSSF: XSSFConditionalFormattingRule, XSSFDxfStyleProvider, XSSFOptimiser,
  XSSFChildAnchor, XSSFFontFormatting, XSSFDataBarFormatting, XSSFColor, XSSFTableStyle,
  XSSFWorkbook, XSSFSimpleShape, XSSFName, ColumnHelper
- poi-ooxml XSLF: XSLFSimpleShape, XSLFSheet, XSLFTabStop, XSLFDrawing
- poi-ooxml XDDF: XDDFDataSourcesFactory
- hwpf: PlfLfo, RecordUtil
- util: IOUtils
- poifs/nio: ByteArrayBackedDataSource

Guarded casts (already behind Integer.MAX_VALUE checks) and casts of
floating-point or char values are intentionally left unchanged.



* Fix test failures: revert (int) cast for EMF uint32 identifiers and palette/brush indices, fix PropertySetFactory for large sectionCount

Agent-Logs-Url: https://github.com/pjfanning/poi/sessions/1fcc029e-f25e-4e46-be29-fc0831ec0ee1


* Address review comments: use Math.toIntExact with try/catch in PropertySetFactory, keep (Long) cast in XSSFChildAnchor

Agent-Logs-Url: https://github.com/pjfanning/poi/sessions/1fcc029e-f25e-4e46-be29-fc0831ec0ee1


* exceptions

* Update poi-integration-exceptions.csv

* Update TestAllFiles.java

---------

+++ poi-ooxml/src/main/java/org/apache/poi/xssf/usermodel/XSSFDataBarFormatting.java
@@ -52,14 +52,14 @@ public class XSSFDataBarFormatting implements DataBarFormatting {
     }
 
     public int getWidthMin() {
-        return (int)_databar.getMinLength();
+        return Math.toIntExact(_databar.getMinLength());
     }
     public void setWidthMin(int width) {
         _databar.setMinLength(width);
     }
 
     public int getWidthMax() {
-        return (int)_databar.getMaxLength();
+        return Math.toIntExact(_databar.getMaxLength());
     }
     public void setWidthMax(int width) {
         _databar.setMaxLength(width);
